+++ src/services/api.js
@@ -1,12 +1,10 @@
 export async function getCategories() {
-  // Implemente aqui
   const response = await fetch('https://api.mercadolibre.com/sites/MLB/categories');
   const categories = await response.json();
   return categories;
 }
 
 export async function getProductsFromCategoryAndQuery(query, categoryId = '') {
-  // Implemente aqui! Quando o fizer, descomente os parâmetros que essa função recebe
   const response = await fetch(`https://api.mercadolibre.com/sites/MLB/search?category=${categoryId}&q=${query}`);
   const product = await response.json();
   return product;
